requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.1
Flox-lib==0.20.1
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==4.9.3
pyperclip==1.8.2
requests==2.31.0
Reverso-API==0.0.1b3
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.5
urllib3==2.0.7
